machop/requirements.txt
pytorch-lightning
transformers
toml
timm
pytorch-nlp
datasets
onnx
black
GitPython
colorlog
cocotb[bus]==1.8.0
pytest
ipython
ipdb
sentencepiece
einops
deepspeed
pybind11
tabulate
tensorboard
hyperopt
